Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I think I've seen patches like these before, would be great if you could maintain authorship unless you're really the original author

Well, your're right. I'm not the original author. But I wasn't able to maintain the authorship from this commit: https://gerrit.omnirom.org/#/c/28163/ because the filename of msm8974.mk respectively msm8974_omni.mk changed. My git knowledge is to limited for this case and I was too lazy for searching hours. I thought a link in the commit message may be enough. Is a cherry-pick the right approach? Advice is welcome ;)

Patch-set: 1