Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review+1
> > 
> > Tested and works great. Unfortunately it's illegal in a lot of countries.
> > Lineage solved this with rules for every country (function depends on MCC)
> > I temporary disabled this MCC https://review.lineageos.org/#/c/LineageOS/android_packages_apps_Dialer/+/226101/3/java/com/android/dialer/callrecord/res/values-mcc262/config.xml therefore.
> > 
> > Can you test this on Castor?
> 
> What do you mean with "temporary disabled"?
> Can't test on castor_windy, doesn't have SIM/phone functionality

Call recording doesn't work in some countries due to legal issues. That's also the case in germany. So I wasn't able to verify the function.
The dialer checks the MCC (reported by SIM) to determine the country.
Then it looks up the MCC if it's legal to record a call in that country.
Now I just "disabled" the country blocker for germany ONLY for testing/verification purposes.
(wanted to see if it actually works.)

Patch-set: 1